// src/BFF/Private/p007/dto/P007Filtros.dto.ts
import { CampeonatoEnum } from 'src/Constantes/CampeonatoEnum';
import { EstiloEnum } from 'src/Constantes/EstiloEnum';
import { GeneroEnum } from 'src/Constantes/GeneroEnum';
import { PiscinaEnum } from 'src/Constantes/PiscinaEnum';
import { PruebaEnum } from 'src/Constantes/PruebaEnum';
import { TemporadaEnum } from 'src/Constantes/TemporadaEnum';

export class P007FiltrosDto {
  year?: number;
  temporada?: TemporadaEnum;
  piscina?: PiscinaEnum;
  estilo?: EstiloEnum;
  prueba?: PruebaEnum;
  genero?: GeneroEnum;
  categoria?: number;
  campeonato?: CampeonatoEnum;
  IDNadador?: number;
}